refactor(core): improve code around element validation
improve code regarding element validation by creating a new file
exporting validation functions and not exporting utils previously
globally available
